Money Class and Currency class
Money class is partially done. Please review the changes and merge accordingly. i have changed some return types so merge the changes with care.

Currency class. I have made change in the "valueInThisCurrency" method. i changed the rate as i thought that was necessary as we need to add the current rate and not hardcode it with 0.27. please review and merge accordingly.